+++ valhalla/baldr/graphtileheader.h
@@ -18,7 +18,7 @@ namespace baldr {
 // something to the tile simply subtract one from this number and add it
 // just before the empty_slots_ array below. NOTE that it can ONLY be an
 // offset in bytes and NOT a bitfield or union or anything of that sort
-constexpr size_t kEmptySlots = 11;
+constexpr size_t kEmptySlots = 10;
 
 // Maximum size of the version string (stored as a fixed size
 // character array so the GraphTileHeader size remains fixed).
@@ -246,6 +246,10 @@ public:
       LOG_ERROR("Tile exceeded maximum directededge count: " + std::to_string(count));
     }
     directededgecount_ = count;
+  }
+
+  bool has_bounding_circles() const {
+    return boundingcircles_offset_ != 0 && boundingcircles_offset_ != tile_size_;
   }
 
   /**
@@ -585,6 +589,22 @@ public:
   }
 
   /**
+   * Get the offset to the start of the bounding circles
+   * @return the byte offset to the start of the bounding circles
+   */
+  uint32_t bounding_circle_offset() const {
+    return boundingcircles_offset_ == tile_size_ ? 0 : boundingcircles_offset_;
+  }
+
+  /**
+   * Sets the offset to the start of the bounding circles
+   * @param offset the offset in bytes to the beginning of the bounding circles
+   */
+  void set_bounding_circle_offset(uint32_t offset) {
+    boundingcircles_offset_ = offset;
+  }
+
+  /**
    * Get the checksum hash of the tile
    * @return return the 64bit hash of tile's input checksum
    */
@@ -702,6 +722,7 @@ protected:
   // GraphTile data size in bytes
   uint32_t tile_size_ = 0;
 
+  uint32_t boundingcircles_offset_ = 0;
   // Marks the end of this version of the tile with the rest of the slots
   // being available for growth. If you want to use one of the empty slots,
   // simply add a uint32_t some_offset_; just above empty_slots_ and decrease